Refactor disassembler: introduce layout system, address attributes, and data types
Move disassembly logic (disassemble, generate_auto_labels, mark_data_attributes)
into Project methods, replacing the ad-hoc listing pipeline in disasm.rs with a
LayoutBuilder/Widgets rendering system backed by new layout.rs,
address_attributes.rs, and data_type.rs modules.

- Delete attributes.rs and listing_line.rs; replace with address_attributes.rs
- Rename imm_seg → ofs_seg throughout (DisplayContext, project files, cpu_impl)
- Broaden address offsets from u16 to u32 in BranchMap and related APIs
- Split DecodedInstruction::format into format_opcode + format_arg for finer control
- Add struct definitions to .chani project format (e.g. IntroPart in cryo-dune)
- Add smallstr dependency for small-string optimization

+++ crates/chani_disasm/src/branch_map.rs
@@ -1,8 +1,8 @@
 use std::collections::{BTreeMap, BTreeSet};
 
-pub type Addr = (usize, u16); // (segment index, offset)
+pub type Addr = (usize, u32); // (segment index, offset)
 
-#[derive(Debug, Default)]
+#[derive(Debug, Default, Clone)]
 pub struct BranchMap {
     forward: BTreeMap<Addr, BTreeSet<Addr>>,
     backward: BTreeMap<Addr, BTreeSet<Addr>>,
